+++ CHANGES.md
@@ -1,5 +1,11 @@
 # Changelog for Aurelix
 
+## 0.1.1 (unreleased)
+
+- added field level permission filter
+- added encrypted-string field
+- exclude null from response objects
+
 ## 0.1 (2023-08-21)
 
 - initial release